Python 3.6/Authentication_MS/requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
clickclick==1.2.2
connexion==1.3
dnspython==1.15.0
Flask==0.12.2
idna==2.6
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
kafka-python==1.4.3
MarkupSafe==1.0
ply==3.11
py-zipkin==0.12.0
PyJWT==1.6.0
PyYAML==3.12
requests==2.18.4
six==1.11.0
swagger-spec-validator==2.1.0
thriftpy==0.3.9
urllib3==1.22
Werkzeug==0.14.1
